+++ .github/workflows/test.yml
@@ -41,4 +41,4 @@ jobs:
     - name: Run tests
       run: cargo test --all
     - name: Run benchmarking tests
-      run: cargo test --features runtime-benchmarks -p acala-runtime benchmarking
+      run: cargo test --features runtime-benchmarks -p dev-runtime benchmarking
