+++ Cargo.toml
@@ -2,87 +2,38 @@
 name = "acala"
 version = "0.5.4"
 authors = ["Acala Developers"]
-build = "build.rs"
 edition = "2018"
-
-[profile.release]
-panic = "unwind"
 
 [[bin]]
 name = "acala"
 path = "src/main.rs"
 
 [dependencies]
-codec = { package = "parity-scale-codec", version = "1.3.0" }
 futures = "0.3.4"
-hex-literal = "0.2.1"
-jsonrpc-core = "14.0.3"
-log = "0.4.8"
 parity-util-mem = { version = "0.7.0", default-features = false, features = ["jemalloc-global"] }
-serde = { version = "1.0.101", features = ["derive"] }
-serde_json = "1.0.41"
-structopt = "0.3.8"
-
-pallet-transaction-payment-rpc = { version = "2.0.0-rc5" }
-sc-authority-discovery = { version = "0.8.0-rc5" }
-sc-basic-authorship = { version = "0.8.0-rc5" }
-sc-chain-spec = { version = "2.0.0-rc5" }
-sc-cli = { version = "0.8.0-rc5" }
-sc-client-api = { version = "2.0.0-rc5" }
-sc-client-db = { version = "0.8.0-rc5" }
-sc-consensus = { version = "0.8.0-rc5" }
-sc-consensus-babe = { version = "0.8.0-rc5" }
-sc-consensus-babe-rpc = { version = "0.8.0-rc5" }
-sc-consensus-epochs = { version = "0.8.0-rc5" }
-sc-executor = { version = "0.8.0-rc5" }
-sc-finality-grandpa = { version = "0.8.0-rc5" }
-sc-finality-grandpa-rpc = { version = "0.8.0-rc5" }
-sc-keystore = { version = "2.0.0-rc5" }
-sc-network = { version = "0.8.0-rc5" }
-sc-offchain = { version = "2.0.0-rc5" }
-sc-rpc = { version = "2.0.0-rc5" }
-sc-rpc-api = { version = "0.8.0-rc5" }
-sc-service = { version = "0.8.0-rc5" }
-sc-telemetry = { version = "2.0.0-rc5" }
-sc-transaction-pool = { version = "2.0.0-rc5" }
-sp-api = { version = "2.0.0-rc5" }
-sp-block-builder = { version = "2.0.0-rc5" }
-sp-consensus = { version = "0.8.0-rc5" }
-sp-consensus-babe = { version = "0.8.0-rc5" }
-sp-core = { version = "2.0.0-rc5" }
-sp-finality-grandpa = { version = "2.0.0-rc5" }
-sp-inherents = { version = "2.0.0-rc5" }
-sp-io = { version = "2.0.0-rc5" }
-sp-runtime = { version = "2.0.0-rc5" }
-sp-transaction-pool = { version = "2.0.0-rc5" }
-substrate-frame-rpc-system = { version = "2.0.0-rc5" }
-frame-benchmarking-cli = { version = "2.0.0-rc5" }
-sp-blockchain = { version = "2.0.0-rc5" }
-frame-benchmarking = { version = "2.0.0-rc5" }
-
-module-dex-rpc = { path = "modules/dex/rpc" }
-module-support = { path = "modules/support" }
-module-staking-pool-rpc = { path = "modules/staking_pool/rpc" }
-orml-oracle-rpc = { path = "orml/oracle/rpc" }
-orml-utilities = { path = "orml/utilities" }
-runtime = { package = "acala-runtime", path = "runtime" }
-inspect = { package = "acala-inspect", path = "inspect" }
-
-[build-dependencies]
-vergen = "3.0.4"
-build-script-utils = { package = "substrate-build-script-utils", version = "2.0.0-rc5" }
-frame-benchmarking-cli = { version = "2.0.0-rc5" }
+acala-cli = { path = "cli" }
+acala-service = { path = "service" }
 
 [features]
-default = ["wasmtime"]
-wasmtime = [ "sc-cli/wasmtime" ]
 runtime-benchmarks = [
-	"runtime/runtime-benchmarks",
+	"acala-cli/runtime-benchmarks",
 ]
+
+[profile.release]
+# Acala runtime requires unwinding.
+panic = "unwind"
 
 [workspace]
 members = [
 	"modules/*",
+	"cli",
+	"inspect",
+	"primitives",
+	"rpc",
+	"runtime/common",
+	"runtime/dev",
+	"service",
+
 	"orml/auction",
 	"orml/authority",
 	"orml/benchmarking",
@@ -93,12 +44,9 @@ members = [
 	"orml/traits",
 	"orml/utilities",
 	"orml/vesting",
-	"runtime",
-	"inspect",
 	"ecosystem-modules/template/template",
 	"ecosystem-modules/ren/renvm-bridge",
 ]
-exclude = [ "runtime/wasm" ]
 
 # The list of dependencies below (which can be both direct and indirect dependencies) are crates
 # that are suspected to be CPU-intensive, and that are unlikely to require debugging (as some of
