+++ cli/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "acala-cli"
+version = "0.5.4"
+authors = ["Acala Developers"]
+edition = "2018"
+
+[dependencies]
+structopt = "0.3.8"
+sp-core = { version = "2.0.0-rc5", optional = true }
+sc-cli = { version = "0.8.0-rc5", optional = true }
+sc-service = { version = "0.8.0-rc5", optional = true }
+frame-benchmarking-cli = { version = "2.0.0-rc5", optional = true }
+
+inspect = { package = "acala-inspect", path = "../inspect" }
+service = { package = "acala-service", path = "../service", default-features = false, optional = true }
+
+[build-dependencies]
+substrate-build-script-utils = { version = "2.0.0-rc5", default-features = false }
+
+[features]
+default = ["wasmtime", "cli"]
+wasmtime = [ "sc-cli/wasmtime" ]
+cli = [
+	"service",
+	"sp-core",
+	"sc-cli",
+	"sc-service",
+	"frame-benchmarking-cli",
+]
+runtime-benchmarks = [ "service/runtime-benchmarks" ]
