+++ cli/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+	substrate_build_script_utils::generate_cargo_keys();
+}
