+++ cli/src/lib.rs
@@ -1,0 +1,18 @@
+//! Acala CLI library.
+
+#![warn(missing_docs)]
+#![warn(unused_extern_crates)]
+
+#[cfg(feature = "cli")]
+mod cli;
+#[cfg(feature = "cli")]
+mod command;
+
+#[cfg(feature = "cli")]
+pub use cli::*;
+
+#[cfg(feature = "cli")]
+pub use command::*;
+
+#[cfg(feature = "cli")]
+pub use sc_cli::{Error, Result};
