+++ ecosystem-modules/ren/renvm-bridge/Cargo.toml
@@ -13,7 +13,7 @@ sp-std = { version = "2.0.0-rc5", default-features = false }
 sp-io = { version = "2.0.0-rc5", default-features = false }
 sp-runtime = { version = "2.0.0-rc5", default-features = false }
 orml-traits = { path = "../../../orml/traits", default-features = false }
-primitives = { package = "module-primitives", path = "../../../modules/primitives", default-features = false }
+primitives = { package = "acala-primitives", path = "../../../primitives", default-features = false }
 
 [dev-dependencies]
 hex-literal = { version = "0.2.1" }
