+++ modules/airdrop/Cargo.toml
@@ -10,7 +10,7 @@ codec = { package = "parity-scale-codec", version = "1.3.0", default-features = 
 sp-runtime = { version = "2.0.0-rc5", default-features = false }
 frame-support = { version = "2.0.0-rc5", default-features = false }
 frame-system = { version = "2.0.0-rc5", default-features = false }
-primitives = { package = "module-primitives", path = "../primitives", default-features = false }
+primitives = { package = "acala-primitives", path = "../../primitives", default-features = false }
 orml-utilities = { path = "../../orml/utilities", default-features = false }
 
 [dev-dependencies]
