+++ modules/cdp_treasury/Cargo.toml
@@ -15,7 +15,7 @@ orml-traits = { path = "../../orml/traits", default-features = false }
 orml-tokens = { path = "../../orml/tokens", default-features = false }
 orml-utilities = { path = "../../orml/utilities", default-features = false }
 support = { package = "module-support", path = "../support", default-features = false }
-primitives = { package = "module-primitives", path = "../primitives", default-features = false }
+primitives = { package = "acala-primitives", path = "../../primitives", default-features = false }
 frame-benchmarking = { version = "2.0.0-rc5", default-features = false, optional = true }
 
 [dev-dependencies]
