+++ modules/emergency_shutdown/Cargo.toml
@@ -15,7 +15,7 @@ sp-runtime = { version = "2.0.0-rc5", default-features = false }
 sp-std = { version = "2.0.0-rc5", default-features = false }
 support = { package = "module-support", path = "../support", default-features = false }
 loans = { package = "module-loans", path = "../loans", default-features = false }
-primitives = { package = "module-primitives", path = "../primitives", default-features = false }
+primitives = { package = "acala-primitives", path = "../../primitives", default-features = false }
 
 [dev-dependencies]
 sp-core = { version = "2.0.0-rc5", default-features = false }
