+++ modules/homa/Cargo.toml
@@ -12,7 +12,7 @@ frame-support = { version = "2.0.0-rc5", default-features = false }
 frame-system = { version = "2.0.0-rc5", default-features = false }
 staking-pool = { package = "module-staking-pool", path = "../staking_pool", default-features = false }
 support = { package = "module-support", path = "../support", default-features = false }
-primitives = { package = "module-primitives", path = "../primitives", default-features = false }
+primitives = { package = "acala-primitives", path = "../../primitives", default-features = false }
 orml-utilities = { path = "../../orml/utilities", default-features = false }
 
 [features]
