+++ modules/honzon/Cargo.toml
@@ -16,7 +16,7 @@ orml-utilities = { path = "../../orml/utilities", default-features = false }
 cdp-engine = { package = "module-cdp-engine", path = "../cdp_engine", default-features = false }
 loans = { package = "module-loans", path = "../loans", default-features = false }
 support = { package = "module-support", path = "../support", default-features = false }
-primitives = { package = "module-primitives", path = "../primitives", default-features = false }
+primitives = { package = "acala-primitives", path = "../../primitives", default-features = false }
 
 [dev-dependencies]
 sp-io = { version = "2.0.0-rc5", default-features = false }
