+++ modules/polkadot_bridge/Cargo.toml
@@ -14,7 +14,7 @@ orml-utilities = { path = "../../orml/utilities", default-features = false }
 sp-runtime = { version = "2.0.0-rc5", default-features = false }
 sp-std = { version = "2.0.0-rc5", default-features = false }
 support = { package = "module-support", path = "../support", default-features = false }
-primitives = { package = "module-primitives", path = "../primitives", default-features = false }
+primitives = { package = "acala-primitives", path = "../../primitives", default-features = false }
 
 [dev-dependencies]
 sp-core = { version = "2.0.0-rc5", default-features = false }
