--- modules/primitives/src/lib.rs
@@ -1,36 +1,0 @@
-#![cfg_attr(not(feature = "std"), no_std)]
-
-use codec::{Decode, Encode};
-use sp_runtime::RuntimeDebug;
-
-#[cfg(feature = "std")]
-use serde::{Deserialize, Serialize};
-
-#[derive(Encode, Decode, Eq, PartialEq, Copy, Clone, RuntimeDebug, PartialOrd, Ord)]
-#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
-pub enum CurrencyId {
-	ACA = 0,
-	AUSD = 1,
-	DOT = 2,
-	XBTC = 3,
-	LDOT = 4,
-	RENBTC = 5,
-}
-
-#[derive(Encode, Decode, Eq, PartialEq, Copy, Clone, RuntimeDebug, PartialOrd, Ord)]
-#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
-pub enum AirDropCurrencyId {
-	KAR = 0,
-	ACA,
-}
-
-/// Counter for the number of eras that have passed.
-pub type EraIndex = u32;
-
-/// Balance of an account.
-pub type Balance = u128;
-
-/// Signed version of Balance
-pub type Amount = i128;
-
-pub type AuctionId = u32;
