+++ primitives/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "module-primitives"
+name = "acala-primitives"
 version = "0.5.4"
 authors = ["Acala Developers"]
 edition = "2018"
@@ -7,6 +7,7 @@ edition = "2018"
 [dependencies]
 serde = { version = "1.0.101", optional = true }
 codec = { package = "parity-scale-codec", version = "1.3.0", default-features = false }
+sp-core = { version = "2.0.0-rc5", default-features = false }
 sp-runtime = { version = "2.0.0-rc5", default-features = false }
 
 [features]
@@ -15,4 +16,5 @@ std = [
 	"serde",
 	"codec/std",
 	"sp-runtime/std",
+	"sp-core/std",
 ]
