+++ rpc/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "acala-rpc"
+version = "0.5.4"
+authors = ["Acala Developers"]
+edition = "2018"
+
+[dependencies]
+codec = { package = "parity-scale-codec", version = "1.3.0" }
+jsonrpc-core = "14.0.3"
+primitives = { package = "acala-primitives", path = "../primitives" }
+sp-api = { version = "2.0.0-rc5" }
+sp-blockchain = { version = "2.0.0-rc5" }
+sp-block-builder = { version = "2.0.0-rc5" }
+sp-consensus = { version = "0.8.0-rc5" }
+sp-consensus-babe = { version = "0.8.0-rc5" }
+sp-transaction-pool = { version = "2.0.0-rc5" }
+sc-client-api = { version = "2.0.0-rc5" }
+sc-consensus-babe = { version = "0.8.0-rc5" }
+sc-consensus-babe-rpc = { version = "0.8.0-rc5" }
+sc-consensus-epochs = { version = "0.8.0-rc5" }
+sc-finality-grandpa = { version = "0.8.0-rc5" }
+sc-finality-grandpa-rpc = { version = "0.8.0-rc5" }
+sc-keystore = { version = "2.0.0-rc5" }
+sc-rpc = { version = "2.0.0-rc5" }
+substrate-frame-rpc-system = { version = "2.0.0-rc5" }
+pallet-transaction-payment-rpc = { version = "2.0.0-rc5" }
+dev-runtime = { path = "../runtime/dev" }
+module-staking-pool-rpc = { path = "../modules/staking_pool/rpc" }
+orml-oracle-rpc = { path = "../orml/oracle/rpc" }
