+++ runtime/dev/tests/integration_test.rs
@@ -1,7 +1,7 @@
 #![cfg(test)]
 
-use acala_runtime::{
-	get_all_module_accounts, AccountId, Balance, CurrencyId, GetNativeCurrencyId, NewAccountDeposit, Runtime,
+use dev_runtime::{
+	get_all_module_accounts, AccountId, Balance, CurrencyId, Event, GetNativeCurrencyId, NewAccountDeposit, Runtime,
 };
 use frame_support::{
 	assert_noop, assert_ok,
@@ -111,21 +111,21 @@ pub fn origin_of(account_id: AccountId) -> <Runtime as frame_system::Trait>::Ori
 fn set_oracle_price(prices: Vec<(CurrencyId, Price)>) -> DispatchResult {
 	OracleModule::on_finalize(0);
 	assert_ok!(OracleModule::feed_values(
-		<acala_runtime::Runtime as frame_system::Trait>::Origin::none(),
+		<Runtime as frame_system::Trait>::Origin::none(),
 		prices.clone(),
 		0,
 		0,
 		Default::default()
 	));
 	assert_ok!(OracleModule::feed_values(
-		<acala_runtime::Runtime as frame_system::Trait>::Origin::none(),
+		<Runtime as frame_system::Trait>::Origin::none(),
 		prices.clone(),
 		1,
 		0,
 		Default::default()
 	));
 	assert_ok!(OracleModule::feed_values(
-		<acala_runtime::Runtime as frame_system::Trait>::Origin::none(),
+		<Runtime as frame_system::Trait>::Origin::none(),
 		prices,
 		2,
 		0,
@@ -184,10 +184,10 @@ fn emergency_shutdown_and_cdp_treasury() {
 				module_emergency_shutdown::Error::<Runtime>::CanNotRefund,
 			);
 			assert_ok!(EmergencyShutdownModule::emergency_shutdown(
-				<acala_runtime::Runtime as frame_system::Trait>::Origin::root()
+				<Runtime as frame_system::Trait>::Origin::root()
 			));
 			assert_ok!(EmergencyShutdownModule::open_collateral_refund(
-				<acala_runtime::Runtime as frame_system::Trait>::Origin::root()
+				<Runtime as frame_system::Trait>::Origin::root()
 			));
 			assert_ok!(EmergencyShutdownModule::refund_collaterals(
 				origin_of(AccountId::from(ALICE)),
@@ -250,7 +250,7 @@ fn liquidate_cdp() {
 			));
 
 			assert_ok!(CdpEngineModule::set_collateral_params(
-				<acala_runtime::Runtime as frame_system::Trait>::Origin::root(),
+				<Runtime as frame_system::Trait>::Origin::root(),
 				CurrencyId::XBTC,
 				Change::NewValue(Some(Rate::zero())),
 				Change::NewValue(Some(Ratio::saturating_from_rational(200, 100))),
@@ -293,7 +293,7 @@ fn liquidate_cdp() {
 			assert_eq!(AuctionManagerModule::collateral_auctions(0), None);
 
 			assert_ok!(CdpEngineModule::set_collateral_params(
-				<acala_runtime::Runtime as frame_system::Trait>::Origin::root(),
+				<Runtime as frame_system::Trait>::Origin::root(),
 				CurrencyId::XBTC,
 				Change::NoChange,
 				Change::NewValue(Some(Ratio::saturating_from_rational(400, 100))),
@@ -308,7 +308,7 @@ fn liquidate_cdp() {
 			));
 
 			let liquidate_alice_xbtc_cdp_event =
-				acala_runtime::Event::module_cdp_engine(module_cdp_engine::RawEvent::LiquidateUnsafeCDP(
+				Event::module_cdp_engine(module_cdp_engine::RawEvent::LiquidateUnsafeCDP(
 					CurrencyId::XBTC,
 					AccountId::from(ALICE),
 					amount(10),
@@ -336,7 +336,7 @@ fn liquidate_cdp() {
 			));
 
 			let liquidate_bob_xbtc_cdp_event =
-				acala_runtime::Event::module_cdp_engine(module_cdp_engine::RawEvent::LiquidateUnsafeCDP(
+				Event::module_cdp_engine(module_cdp_engine::RawEvent::LiquidateUnsafeCDP(
 					CurrencyId::XBTC,
 					AccountId::from(BOB),
 					amount(1),
@@ -394,7 +394,7 @@ fn test_dex_module() {
 				10000000
 			));
 
-			let add_liquidity_event = acala_runtime::Event::module_dex(module_dex::RawEvent::AddLiquidity(
+			let add_liquidity_event = Event::module_dex(module_dex::RawEvent::AddLiquidity(
 				AccountId::from(ALICE),
 				CurrencyId::XBTC,
 				10000,
@@ -467,7 +467,7 @@ fn test_honzon_module() {
 			)]));
 
 			assert_ok!(CdpEngineModule::set_collateral_params(
-				<acala_runtime::Runtime as frame_system::Trait>::Origin::root(),
+				<Runtime as frame_system::Trait>::Origin::root(),
 				CurrencyId::XBTC,
 				Change::NewValue(Some(Rate::saturating_from_rational(1, 100000))),
 				Change::NewValue(Some(Ratio::saturating_from_rational(3, 2))),
@@ -507,7 +507,7 @@ fn test_honzon_module() {
 				false
 			);
 			assert_ok!(CdpEngineModule::set_collateral_params(
-				<acala_runtime::Runtime as frame_system::Trait>::Origin::root(),
+				<Runtime as frame_system::Trait>::Origin::root(),
 				CurrencyId::XBTC,
 				Change::NoChange,
 				Change::NewValue(Some(Ratio::saturating_from_rational(3, 1))),
@@ -556,7 +556,7 @@ fn test_cdp_engine_module() {
 		.execute_with(|| {
 			SystemModule::set_block_number(1);
 			assert_ok!(CdpEngineModule::set_collateral_params(
-				<acala_runtime::Runtime as frame_system::Trait>::Origin::root(),
+				<Runtime as frame_system::Trait>::Origin::root(),
 				CurrencyId::XBTC,
 				Change::NewValue(Some(Rate::saturating_from_rational(1, 100000))),
 				Change::NewValue(Some(Ratio::saturating_from_rational(3, 2))),
@@ -647,9 +647,10 @@ fn test_cdp_engine_module() {
 				CurrencyId::XBTC
 			));
 
-			let settle_cdp_in_debit_event = acala_runtime::Event::module_cdp_engine(
-				module_cdp_engine::RawEvent::SettleCDPInDebit(CurrencyId::XBTC, AccountId::from(ALICE)),
-			);
+			let settle_cdp_in_debit_event = Event::module_cdp_engine(module_cdp_engine::RawEvent::SettleCDPInDebit(
+				CurrencyId::XBTC,
+				AccountId::from(ALICE),
+			));
 			assert!(SystemModule::events()
 				.iter()
 				.any(|record| record.event == settle_cdp_in_debit_event));
