+++ service/Cargo.toml
@@ -1,0 +1,55 @@
+[package]
+name = "acala-service"
+version = "0.5.4"
+authors = ["Acala Developers"]
+edition = "2018"
+
+[dependencies]
+hex-literal = "0.2.1"
+serde = { version = "1.0.102", features = ["derive"] }
+serde_json = "1.0.41"
+codec = { package = "parity-scale-codec", version = "1.3.0" }
+sp-api = { version = "2.0.0-rc5" }
+sp-consensus = { version = "0.8.0-rc5" }
+sp-consensus-babe = { version = "0.8.0-rc5" }
+sp-core = { version = "2.0.0-rc5" }
+sp-finality-grandpa = { version = "2.0.0-rc5" }
+sp-runtime = { version = "2.0.0-rc5" }
+sp-inherents = { version = "2.0.0-rc5" }
+sc-finality-grandpa = { version = "0.8.0-rc5" }
+sc-chain-spec = { version = "2.0.0-rc5" }
+sc-client-api = { version = "2.0.0-rc5" }
+sc-telemetry = { version = "2.0.0-rc5" }
+sc-service = { version = "0.8.0-rc5" }
+sc-executor = { version = "0.8.0-rc5" }
+sc-consensus = { version = "0.8.0-rc5" }
+sc-consensus-babe = { version = "0.8.0-rc5" }
+sc-transaction-pool = { version = "2.0.0-rc5" }
+sc-basic-authorship = { version = "0.8.0-rc5" }
+sc-network = { version = "0.8.0-rc5" }
+frame-benchmarking = { version = "2.0.0-rc5" }
+
+sp-block-builder = { version = "2.0.0-rc5" }
+frame-system-rpc-runtime-api = { version = "2.0.0-rc5" }
+pallet-transaction-payment-rpc-runtime-api = { version = "2.0.0-rc5" }
+sp-offchain = { version = "2.0.0-rc5" }
+sp-session = { version = "2.0.0-rc5" }
+sp-transaction-pool = { version = "2.0.0-rc5" }
+sp-blockchain = { version = "2.0.0-rc5" }
+
+module-staking-pool-rpc = { path = "../modules/staking_pool/rpc" }
+orml-oracle-rpc = { path = "../orml/oracle/rpc" }
+acala-primitives = { path = "../primitives" }
+acala-rpc = { path = "../rpc" }
+dev-runtime = { path = "../runtime/dev" }
+runtime-common = { path = "../runtime/common" }
+
+[dev-dependencies]
+sc-consensus-babe = { version = "0.8.0-rc5", features = ["test-helpers"] }
+
+[features]
+default = ["std"]
+std = []
+runtime-benchmarks = [
+	"dev-runtime/runtime-benchmarks",
+]
